Added pronouns to Shift Lead Report. Moved report from Slot.php to ShiftLeadReport.php.
